feat: implement GraphQL mutation resolvers and input DTOs
- Add 6 mutation resolvers (Article, User, Category, Tag, Media, PublicationIssue)
- Add 11 input DTOs for create/update operations
- Implement all mutations from GraphQL schema:
  * Article: create, update, delete, publish, archive, reject, restore, toggleFeatured, tag operations
  * User: create, update, delete, changePassword with validation
  * Category/Tag: create, update, delete with auto-slug generation
  * Media: upload with auth, delete
  * PublicationIssue: create, update, delete with publishedAt handling
- Follow service-layer delegation pattern
- Add proper ID parsing and error handling
- Include authentication context for author/uploader fields

+++ src/main/java/com/uep/pillar/service/StatisticsService.java
@@ -28,7 +28,7 @@ public class StatisticsService {
         long totalViews = articleRepository.sumViewCount();
 
         float avgViewsPerArticle = totalArticles > 0
-                ? (float) totalViews / totalArticles
+                ? (float) totalViews / (float) totalArticles
                 : 0f;
 
         return ArticleStatistics.builder()
